created separate setup file for vitest
